refactor: slim AgentController to just runQuery(), loop stays in index.ts
Per review feedback: AgentController should own just one action — running
a single query turn against the Claude SDK. Setup, command registration,
and the routing loop belong in index.ts (main()) since that's where input
is received and routed to controllers.

- AgentController now has only runQuery() and its private helpers
  (handleAskUserQuestion, handleToolPermission)
- logFull and createFetchModelsFn are exported standalone helpers in
  agent-controller.ts (used by index.ts/main())
- main() in index.ts handles worker mode setup, workspace creation,
  startup banner, command registration, and the routing loop:
    command        → registered command handler
    user prompt    → AgentController.runQuery (via runQueryFn)
    foreman signal → WorkerSession
- index.ts startup block constructs objects and calls main()
- Tests updated: worker.main.test.ts again injects runQueryFn into main();
  repl.query.test.ts tests AgentController.runQuery directly
